Make the test loop a subclass of the BaseEventLoop from asyncio.
Do this now to avoid reimplementing a large portion of the asyncio event
loop.  We can hook in our own advance call where the BaseEventLoop wants
to "run once" and get what should be desired behavior.

Later on we can directly reimplement more if it's needed.

We'll need to do something about the networking calls now on the test
event loop.