Improve (or outright make work) the coro TestCase API.
advance() has bee updated to accept a kwarg that causes the scheduler to
not mark Tasks as ready if they're exactly on the wall clock.  This lets
us check infinitely close right up before the wall clock, and on the wall
clock, to see if a coro runs exactly at a given time.